requirements.txt
wheel
setuptools
build
fastapi
uvicorn
python-dotenv
pdf2image
fastapi-limiter
redis
slowapi
python-multipart
Pillow
PyMuPDF
docx2pdf==0.1.8
numpy
python-barcode
qrcode
